Refactor component into single file
