--- frontend/src/components/Navbar.tsx
@@ -1,12 +1,0 @@
-import "~/components/Navbar.css";
-
-export default function NavBar() {
-  return (
-    <div class="tabsPanel">
-      <img src="/img/logo.svg" alt="FastSearch logo" />
-      <div class="tabs">
-        <a href="/writeup/scope.html">PROJECT WRITEUP</a>
-      </div>
-    </div>
-  );
-}
